Refactor Algolia search bar to separate component
